Adds summary panel
