+++ main/core/Resources/modules/user/tracking/components/main.jsx
@@ -3,65 +3,15 @@ import {PropTypes as T} from 'prop-types'
 import {connect} from 'react-redux'
 import {merge} from 'lodash'
 
-import {CallbackButton} from '#/main/app/buttons/callback/components/button'
-
 import {trans} from '#/main/core/translation'
 import {User} from '#/main/core/user/prop-types'
+import {actions} from '#/main/core/user/tracking/store'
 import {ResourceUserEvaluation} from '#/main/core/user/tracking/prop-types'
 import {UserPageContainer} from '#/main/core/user/containers/page'
 import {UserDetails} from '#/main/core/user/components/details'
 import {Timeline} from '#/main/core/user/tracking/components/timeline'
-import {DateGroup} from '#/main/core/layout/form/components/group/date-group'
-import {actions} from '#/main/core/user/tracking/store'
-
-const Search = props =>
-  <div className="panel panel-default">
-    <div className="panel-body">
-      <div className="col-md-6 col-xs-12">
-        <DateGroup
-            id="tracking-start-date"
-            className="form-last"
-            calendarIcon="fa fa fa-fw fa-calendar-check-o"
-            label={trans('filter_from')}
-            value={props.startDate}
-            onChange={(date) => props.onChange('startDate', date)}
-        />
-      </div>
-      <div className="col-md-6 col-xs-12">
-        <DateGroup
-            id="tracking-end-date"
-            className="form-last"
-            calendarIcon="fa fa fa-fw fa-calendar-check-o"
-            label={trans('date_range_end')}
-            value={props.endDate}
-            minDate={props.startDate}
-            onChange={(date) => props.onChange('endDate', date)}
-        />
-      </div>
-      <div className="col-md-6 col-xs-12">
-        <CallbackButton
-          className="btn btn-primary traking-filter-button"
-          callback={() => props.onSearch(props.startDate, props.endDate)}
-        >
-          {trans('filter')}
-        </CallbackButton>
-      </div>
-    </div>
-  </div>
-
-Search.propTypes = {
-  startDate: T.string,
-  endDate: T.string,
-  onChange: T.func.isRequired,
-  onSearch: T.func.isRequired
-}
-
-const Summary = props =>
-  <div className="panel panel-default">
-    <div className="panel-body">
-      Coucou
-    </div>
-  </div>
+import {Search} from '#/main/core/user/tracking/components/search'
+import {Summary} from '#/main/core/user/tracking/components/summary'
 
 class TrackingComponent extends Component {
   constructor(props) {
@@ -115,7 +65,9 @@ class TrackingComponent extends Component {
               onSearch={this.props.loadTrackings}
             />
 
-            <Summary />
+            <Summary
+              evaluations={this.props.evaluations}
+            />
 
             <Timeline
               events={this.props.evaluations.map(e => {return {
